[fix] 코드래빗 수정사항 반영

+++ src/main/java/kr/gilmok/api/reservation/dto/ReservationStatsResponse.java
@@ -5,7 +5,7 @@ public record ReservationStatsResponse(
         long holdingCount,
         long confirmedCount,
         long cancelledCount,
-        int holdingQuantity,
-        int confirmedQuantity,
-        int cancelledQuantity
+        long holdingQuantity,
+        long confirmedQuantity,
+        long cancelledQuantity
 ) {}
